refactor: 포맷된 입력 필드 로직 재사용 가능하도록 리팩토링 (#95)
- FormattedInput 컴포넌트 추출
- 포맷팅 함수를 utils/formatter.ts로 분리
- BasicInfo 컴포넌트에서 중복 코드 제거
